Better way to deal with all async stuff
